Added ManagedProcess class
ManagedProcess is designed to manage the execution of a particular
process. When a process is started, a new IOManager will be created to
manage communication with the process. ManagedProcess will remember the process location and arguments so that it can be easily started/stopped/restarted as desired using the processbuilder API.