IOManager and ManagedProcess have been refactored
IOManager's input stream has been changed to a bufferedreader. This
allows for realtime reading while the process is running. ManagedProcess
has been modified to start the process in the constructor. This is
because it is possible with the older version to have a process that has
not yet been started and yet access it's IOManager, resulting in a null
pointer exception. More refactoring may be required in the future to
ensure the object cannot have an invalid state.