ManagedProcess refactored
Multi-argument constructor has been reworked to include the "varargs"
syntax as it's much more clean when executing. Additionally, start() has
been moved out of the constructors because of a planned future change
(scheduled program running). If a program is scheduled it should not
immediately start execution, so calling start() in the constructor isn't
correct.