Process status monitoring added to ManagedProcess
ManagedProcess will now create a new thread that will monitor the status
of a process, automatically detecting when the process stops running for
whatever reason.

The IOManager associated with the process will no
longer be automatically destroyed on stop(), as it may be desired to
read the remaining output from the buffer before starting a new process.