Auto restart has been implemented in ManagedProcess
If auto restart is enabled, a process will be immediately restarted upon
exit. This will require some kind of rate limiting in the future, as a
network downloading script shouldn't be continually enabled upon
failure. Perhaps an error check would resolve this issue, but errors
have not been implemented in IOManager yet.