ProcessLogger created and logging system added
ProcessLogger will write either to stdout or a log file. ProcessLogger
is intended to handle the task of writing logs for the process
monitoring thread. Timestamps are enabled by default but may be disabled
with the property "timestamp". By default, logs are generated in the
same directory as the executable, but this may be changed by providing a
path.

ManagedProcess will now actively read from the output of the process on
it's separate thread and feed whatever output it can read to the
processlogger. By default, these log messages will go to stdout and
stderr unless logfile is enabled.