Scheduled task has been added
Scheduled task is essentially a custom-made timer designed to be stored
in a ManagedProcess instance. Upon elapse, this timer will send a signal
to the running process or start/stop/restart the process. This will be
used in the implementation of a scheduling system for ManagedProcess in
the future.

ProcessSignal has been replaced by String for the time being, as it has
proven to not be very useful in actual terms. More often than not, it's
more useful to use a static string than it is to create a signal and
define custom arguments. ProcessSignal will likely be removed in the
future.