ScheduledTask has been updated
The payload of the task is now set via an enum rather than a series of
boolean values, since the elapsed() will be handled by a parent thread.
Time setting functions have been reworked to allow easy setting for a
particular time of day or a specific datetime. Enabled field has been
added to see if a task's payload should be acted upon in the parent
thread or not. Various cleanup has been performed.