General cleanup of ManagedProcess class
Logging is now enabled and not able to be disabled. log messages must be
written to stdout or a file, as multiple instances of ManagedProcess can
(and are intended to be) running simultaneously. Thus their messages
should be handled by a ProcessLogger as it identifies which
ManagedProcess is sending the message.