Maven build settings have been updated
To be honest, I'm not at all familiar with java build systems or what
the best practices are. I'm kind of just taking a guess here. I updated
the groupid to be more in line with recommendations online.
Additionally, the ability to build a jar from maven has (hopefully) been
added.