AuthenticationManager has been added
AuthenticationManager will provide functions for the
authentication of passwords and storage of password hashes. These
password hashes will be generated using PBKDF2 with HMAC SHA256, which
at the time of writing is standard practice for password hashing. This
code should NOT be considered secure, do not depend on it for system
security, as it almost certainly contains mistakes.